opencode_pytest_harness/requirements.txt
pexpect
pytest
pyyaml
ruff
